Post manager complete

+++ controllers/PostController.php
@@ -1,10 +1,14 @@
-<?php
+<?php /** @noinspection PhpUndefinedClassInspection */
 
 namespace app\controllers;
 
+use Throwable;
 use Yii;
 use app\models\Post;
 use app\models\PostSearch;
+use yii\db\StaleObjectException;
+use yii\filters\AccessControl;
+use yii\web\ConflictHttpException;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -20,6 +24,21 @@ class PostController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'rules' => [
+                    [
+                        'actions' => ['index', 'view'],
+                        'allow' => true,
+                        'roles' => ['?'],
+                    ],
+                    [
+                        'actions' => [],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -39,6 +58,21 @@ class PostController extends Controller
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
         return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);
+    }
+
+    /**
+     * Lists all Post models.
+     * @return mixed
+     */
+    public function actionAdmin()
+    {
+        $searchModel = new PostSearch();
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+
+        return $this->render('admin', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
@@ -104,24 +138,48 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (Yii::$app->request->isPost) {
+            // we only allow deletion via POST request
+            try {
+                $this->findModel($id)->delete();
+            } catch (StaleObjectException $e) {
+            } catch (NotFoundHttpException $e) {
+            } catch (Throwable $e) {
+            }
+            if (!Yii::$app->request->isAjax)
+                $this->redirect(array('admin'));
+        } else
+            throw new NotFoundHttpException('Invalid request. Please do not repeat this request again.',404);
+        return $this->redirect(array('admin'));
+//        return $this->redirect(['view', 'id' => $id]);
     }
+
+    private $_model;
 
     /**
      * Finds the Post model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
      * @param integer $id
-     * @return Post the loaded model
+     * @return array|\yii\db\ActiveRecord
      * @throws NotFoundHttpException if the model cannot be found
      */
     protected function findModel($id)
     {
-        if (($model = Post::findOne($id)) !== null) {
-            return $model;
+        if ($this->_model === null) {
+            if (isset($id)) {
+                if (Yii::$app->user->isGuest) {
+                    $condition = 'status='
+                        . Post::STATUS_PUBLISHED
+                        . ' OR status='
+                        . Post::STATUS_ARCHIVED;
+                } else {
+                    $condition = '';
+                }
+                $this->_model = Post::find()->where(['id' => $id])->onCondition($condition)->one();
+            }
+            if ($this->_model === null)
+                throw new NotFoundHttpException('Запрашиваемая страница не существует');
         }
-
-        throw new NotFoundHttpException('The requested page does not exist.');
+        return $this->_model;
     }
 }
